api/requirements.txt
fastapi
uvicorn[standard]
kafka-python
pydantic
psycopg2-binary
SQLAlchemy==2.0.32